v1.6.2 - Improved file structure
Added separate py files for the cogs, moved py files into a folder, moved pictures into a folder.
